trying a fix on random(0..n) bug
git-svn-id: svn://danse.us/MCViNE/trunk@1626 bd03dbb6-ca95-436e-b6ad-03a95796df24

+++ packages/mccomponents/libmccomponents/kernels/sample/phonon/CoherentInelastic_SingleXtal.cc
@@ -147,7 +147,7 @@ const
 	<< journal::endl;
 #endif
   // choose a root (length of vf)
-  int index=(int)mccomponents::math::random(0, nf);
+  size_t index=mccomponents::math::random(size_t(0), nf);
   float_t v_f_l = vf_list[index];
 #ifdef DEBUG
   debug << journal::at(__HERE__)
